Pass in familyGuid and add more columns to the splice tab.
